% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{An Adverse Event Dataset}
\format{A data frame with 1191 rows and 55 variables.}
\source{
\url{https://www.cdisc.org/pilot-project-submission-package}
}
\usage{
adae
}
\description{
A dataset containing the adverse event information of a clinical trial following
CDISC ADaM standard.
}
\details{
Definition of each variable can be found in
\url{https://www.cdisc.org/pilot-project-submission-package}
}
\keyword{datasets}
