% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{A Subject Level Demographic Dataset}
\format{
A data frame with 254 rows and 51 variables.
}
\source{
\url{https://bitbucket.cdisc.org/projects/CED/repos/sdtm-adam-pilot-project/browse}
}
\usage{
adsl
}
\description{
A dataset containing the demographic information of a clinical trial following
CDISC ADaM standard.
}
\details{
Definition of each variable can be found in
\url{https://bitbucket.cdisc.org/projects/CED/repos/sdtm-adam-pilot-project/browse}
}
\keyword{datasets}
