% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_colheader}
\alias{as_rtf_colheader}
\title{Create Column Header RTF Encode}
\usage{
as_rtf_colheader(tbl)
}
\arguments{
\item{tbl}{A data frame}
}
\description{
Create Column Header RTF Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Obtain column header attributes from \code{tbl} object.
   \item Extract column header total width from page col_width attribute.
   \item Define column header in RTF syntax using \code{rtf_table_content()}.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

