% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_color}
\alias{as_rtf_color}
\title{Create RTF Color Encode}
\usage{
as_rtf_color(tbl)
}
\arguments{
\item{tbl}{a data frame}
}
\description{
Create RTF Color Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Initiate RTF color using \code{color_table()} if use_color is TRUE in page attribute.
   \item Combine all components into a single RTF code string.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

