% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_end}
\alias{as_rtf_end}
\title{End RTF Encode}
\usage{
as_rtf_end()
}
\description{
End RTF Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Add symbol right curly bracket at the end of code.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

