% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_font}
\alias{as_rtf_font}
\title{Create RTF Font Encode}
\usage{
as_rtf_font()
}
\description{
Create RTF Font Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Initiate RTF font type using \code{font_type()}.
   \item Combine all components into a single RTF code string.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

