% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_footnote}
\alias{as_rtf_footnote}
\title{Create Footnote RTF Encode}
\usage{
as_rtf_footnote(tbl, attr_name = "rtf_footnote")
}
\arguments{
\item{tbl}{A data frame.}

\item{attr_name}{a character value of attribute name.}
}
\description{
Create Footnote RTF Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Obtain footnote attributes from \code{tbl}.
   \item Define footnote in RTF syntax using \code{rtf_table_content()} if as_table attribute is TRUE.
   \item Define footnote in RTF syntax using \code{rtf_paragraph()} if as_table attribute is FALSE.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

