% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_init}
\alias{as_rtf_init}
\title{Create RTF Header Encode}
\usage{
as_rtf_init()
}
\description{
Create RTF Header Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Initiate RTF table by defining language #1033 (U.S. English).
   \item Define the initiation in RTF syntax.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

