% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_margin}
\alias{as_rtf_margin}
\title{Create RTF Page Margin Encode}
\usage{
as_rtf_margin(tbl)
}
\arguments{
\item{tbl}{A data frame.}
}
\description{
Create RTF Page Margin Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect margin attributes from \code{tbl} object.
   \item Convert margin from inch to twip using \code{inch_to_twip()}.
   \item Define margin in RTF syntax.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

