% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_new_page}
\alias{as_rtf_new_page}
\title{Create RTF New Page Encode}
\usage{
as_rtf_new_page()
}
\description{
Create RTF New Page Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define new page in RTF syntax.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

