% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_page}
\alias{as_rtf_page}
\title{Create RTF Page Size Encode}
\usage{
as_rtf_page(tbl)
}
\arguments{
\item{tbl}{a data frame}
}
\description{
Create RTF Page Size Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect page attributes from \code{tbl} object.
   \item Convert page size from inch to twip using \code{inch_to_twip()}.
   \item Define page size in width, height and orientation (landscape or portrait) in RTF syntax.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

