% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_rtf_pageby.R
\name{as_rtf_pageby}
\alias{as_rtf_pageby}
\title{RTF Table Page By Encoding}
\usage{
as_rtf_pageby(tbl)
}
\arguments{
\item{tbl}{A data frame.}
}
\description{
RTF Table Page By Encoding
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect all attributes from \code{tbl} object.
   \item Define table attributes using \code{rtf_table_content()}.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

