% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_source}
\alias{as_rtf_source}
\title{Create Data Source RTF Encode}
\usage{
as_rtf_source(tbl)
}
\arguments{
\item{tbl}{A data frame.}
}
\description{
Create Data Source RTF Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Obtain source attributes from \code{tbl}.
   \item Define source in RTF syntax using \code{rtf_table_content()} if as_table attribute is TRUE.
   \item Define source in RTF syntax using \code{rtf_paragraph()} if as_table attribute is FALSE.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

