% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_subline}
\alias{as_rtf_subline}
\title{Create Table Subline RTF Encode}
\usage{
as_rtf_subline(tbl)
}
\arguments{
\item{tbl}{A data frame.}
}
\description{
Create Table Subline RTF Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Obtain title and subtitle text from \code{tbl} using \code{rtf_text()}.
   \item Define title and subtitle text font, size, format and color attributes.
   \item Return title/subtitle to header using \code{rtf_paragraph()} if not NULL, otherwise return NULL to header.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

