% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_create.R
\name{as_rtf_title}
\alias{as_rtf_title}
\title{Create Table Title RTF Encode}
\usage{
as_rtf_title(tbl)
}
\arguments{
\item{tbl}{A data frame.}
}
\description{
Create Table Title RTF Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Obtain title attributes from \code{tbl} object.
   \item Define title in RTF syntax using \code{as_rtf_paragraph()} if it is not NULL, otherwise return NULL.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

