% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{border_type}
\alias{border_type}
\title{RTF Border Type Dictionary}
\usage{
border_type()
}
\description{
RTF Border Type Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect most commonly used border types for a table.
   \item Define the border types in RTF syntax.
   \item Create a mapping between border types and their RTF code.
   \item Return to `border_type()` data frame to see all available border types.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

