% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cell_size}
\alias{cell_size}
\title{Calculate Cell Size in Twips}
\usage{
cell_size(col_rel_width, col_total_width)
}
\arguments{
\item{col_rel_width}{A vector of numbers separated by comma to indicate column relative width ratio.}

\item{col_total_width}{A numeric number to indicate total column width.}
}
\description{
Calculate Cell Size in Twips
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Convert inch to twip for cell size using \code{.inch_to_twip()}.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

