% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_rich_text.R
\name{check_braces}
\alias{check_braces}
\title{Check that braces are correctly matched}
\usage{
check_braces(input)
}
\arguments{
\item{input}{Plain text containing matched curly braces with tags.}
}
\description{
Braces must be matched appropriately. First brace should be an opening brace.
Every brace should be closed appropriate.
}
\keyword{internal}
