% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{color_table}
\alias{color_table}
\title{RTF Text Color Dictionary}
\usage{
color_table()
}
\description{
RTF Text Color Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect all possible colors from R graphics devices.
   \item Define the colors to RGB conversion in RTF syntax.
   \item Combine all RGB components into a single RTF code string.
   \item Create a mapping between colors and their RTF code.
   \item Return to `color_table()` data frame to see the complete mapping.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

