% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert}
\alias{convert}
\title{Convert Symbol to ANSI and Unicode Encoding}
\usage{
convert(
  text,
  load_stringi = suppressMessages(suppressWarnings(require("stringi")))
)
}
\arguments{
\item{text}{A string to be converted.}

\item{load_stringi}{A logical value to load \code{stringi} or not}
}
\description{
Convert Symbol to ANSI and Unicode Encoding
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define commonly used symbols in RTF syntax (superscript, subscript, greater than or equal to, less than or equal to, line break).
   \item Define Pattern for latex code.
   \item Declare fixed string in the pattern (no regular expression).
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

