% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_rich_text.R
\name{extract_tagged_text}
\alias{extract_tagged_text}
\title{Extract tagged text}
\usage{
extract_tagged_text(input)
}
\arguments{
\item{input}{Plain text containing matched curly braces with tags.}
}
\description{
Identify the text that is in brackets and correctly resolve the ordering of
the brackets such that everything is correctly tagged with the needed style.
}
\keyword{internal}
