% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{fig_format}
\alias{fig_format}
\title{RTF Text Format Dictionary}
\usage{
fig_format()
}
\description{
RTF Text Format Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect supported figure format.
   \item Create a mapping between figure formats and their RTF code.
 }

 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

