% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{font_format}
\alias{font_format}
\title{RTF Text Format Dictionary}
\usage{
font_format()
}
\description{
RTF Text Format Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect most commonly used font formats (normal, bold, italics, underline, strike, superscript, and subscript).
   \item Define font format types in "", "b", "i", "u", "s", "^", "_".
   \item Create a mapping between font formats and their RTF code.
 }

 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

