% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{inch_to_twip}
\alias{inch_to_twip}
\title{Convert Inches to Twips}
\usage{
inch_to_twip(inch)
}
\arguments{
\item{inch}{Value in inches.}
}
\description{
Convert Inches to Twips
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Convert inch to twips using conversion factor 1:1440.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

