% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{justification}
\alias{justification}
\title{RTF Text Justification Dictionary}
\usage{
justification()
}
\description{
RTF Text Justification Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect most commonly used alignments for texts or rows (left, center, right, decimal, and justified).
   \item Define alignments/justifications in "l", "c", "r", "d", "j".
   \item Define the alignments/justifications for texts in RTF syntax.
   \item Define the alignments/justifications for rows in RTF syntax.
   \item Create a mapping between justifications and their RTF code.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

