% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_rich_text.R
\name{match_braces}
\alias{match_braces}
\title{Identify opening and closing brace pairs}
\usage{
match_braces(openings, closings)
}
\arguments{
\item{openings}{Vector of indices indicating location of opening braces.}

\item{closings}{Vector of indices indicating location of closing braces.}
}
\description{
Identify which opening and closing braces in a string belong together. Follows
a first-in-last-out matching.
}
\keyword{internal}
