% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2rtf_adae}
\alias{r2rtf_adae}
\title{An Adverse Event Dataset}
\format{
A data frame with 1191 rows and 55 variables.
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
r2rtf_adae
}
\description{
A dataset containing the adverse event information of a clinical trial following
CDISC ADaM standard.
}
\details{
Definition of each variable can be found in
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\keyword{datasets}
