% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2rtf_HAMD17}
\alias{r2rtf_HAMD17}
\title{An Efficacy Clinical Trial Data to Evaluate a Drug to Reduce Lower Back Pain}
\format{
A data frame with 831 rows and 6 variables.
}
\source{
\url{https://www.lshtm.ac.uk/research/centres-projects-groups/missing-data#dia-missing-data}
}
\usage{
r2rtf_HAMD17
}
\description{
A dataset prepared by the Drug Information Association scientific working group
to investigate a drug to reduce lower back pain.
}
\details{
Definition of each variable can be found in \url{https://www.lshtm.ac.uk/research/centres-projects-groups/missing-data#dia-missing-data}
}
\keyword{datasets}
