% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2rtf_tbl1}
\alias{r2rtf_tbl1}
\title{Within Group Results from an ANCOVA Model}
\format{
A data frame with 2 rows and 8 variables.
}
\usage{
r2rtf_tbl1
}
\description{
A dataset containing within group results from an ANCOVA model.
}
\keyword{datasets}
