% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2rtf_tbl2}
\alias{r2rtf_tbl2}
\title{Between Group Results from an ANCOVA Model}
\format{
A data frame with 1 row and 3 variables.
}
\usage{
r2rtf_tbl2
}
\description{
A dataset containing between group results from an ANCOVA model.
}
\keyword{datasets}
