% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2rtf_tbl3}
\alias{r2rtf_tbl3}
\title{Root Mean Square Error from an ANCOVA model}
\format{
A data frame with 1 row and 1 variable.
}
\usage{
r2rtf_tbl3
}
\description{
A dataset containing root mean square error from an ANCOVA model.
}
\keyword{datasets}
