% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_group_by.R
\name{remove_group_by}
\alias{remove_group_by}
\title{Remove Duplicate Records}
\usage{
remove_group_by(tbl, group_by, page)
}
\arguments{
\item{tbl}{A data frame}

\item{page}{A numeric vector of page index}

\item{by}{A character vector of variable names in tbl}
}
\value{
Return \code{tbl}.
}
\description{
Remove Duplicate Records
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define \code{id} variable to split data frame.
   \item Remove duplicate records within each splitted data frame.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

