% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf2pdf.R
\name{rtf2pdf}
\alias{rtf2pdf}
\title{Convert RTF file to PDF file}
\usage{
rtf2pdf(input, output_file = NULL, output_dir = ".", overwrite = FALSE)
}
\arguments{
\item{input}{A vector of file paths for the input file to be converted.}

\item{output_file}{A vector of filename for the output file. Default is the same filename for input.}

\item{output_dir}{The output directory for the converted \code{output_dir}.}

\item{overwrite}{logical; should existing destination files be overwritten?}
}
\value{
A vector of file paths for the converted files.
}
\description{
Convert RTF files to PDF files. Require libreoffice7.1.
}
