% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_assemble.R
\name{rtf_assemble}
\alias{rtf_assemble}
\title{Assemble RTF TLFs}
\usage{
rtf_assemble(
  input,
  output,
  landscape = FALSE,
  use_officer = require(officer, quietly = TRUE)
)
}
\arguments{
\item{input}{Character vector of file path.}

\item{output}{Character string to the output file path.}

\item{landscape}{Logical vector to determine whether to
display files as portrait or landscape.}

\item{use_officer}{Logical value to determine whether the package officer::
should be used to assemble rtfs.}
}
\description{
Add a set of RTF/TEXT fields into an rdocx object.
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Transfer files to toggle fields format in Word
    \item Insert into Word file using officer
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
\dontrun{
rtf_assemble(
  list.files(
    "outtable/",
    pattern = "*.rtf",
    full.names = TRUE
  ),
  output = "tmp.docx"
)
}

}
