% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_by_subline.R
\name{rtf_by_subline}
\alias{rtf_by_subline}
\title{Add
Sublineby Attributes to Table}
\usage{
rtf_by_subline(tbl, subline_by)
}
\arguments{
\item{tbl}{A data frame.}

\item{subline_by}{Column names in a character vector to subline by table in sections.
Default is NULL.}
}
\description{
Add
Sublineby Attributes to Table
}
