% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_colheader.R
\name{rtf_colheader}
\alias{rtf_colheader}
\title{Add Column Header Attributes to Table}
\usage{
rtf_colheader(
  tbl,
  colheader = NULL,
  col_rel_width = NULL,
  border_left = "single",
  border_right = "single",
  border_top = "single",
  border_bottom = "",
  border_color_left = NULL,
  border_color_right = NULL,
  border_color_top = NULL,
  border_color_bottom = NULL,
  border_width = 15,
  cell_height = 0.15,
  cell_justification = "c",
  text_font = 1,
  text_format = NULL,
  text_font_size = 9,
  text_color = NULL,
  text_background_color = NULL,
  text_justification = "c",
  text_indent_first = 0,
  text_indent_left = 0,
  text_indent_right = 0,
  text_space = 1,
  text_space_before = 15,
  text_space_after = 15,
  text_convert = TRUE
)
}
\arguments{
\item{tbl}{A data frame. Required argument.}

\item{colheader}{A character string that uses " | " to separate column names. Default is NULL for a blank column header.}

\item{col_rel_width}{A numeric vector for column relative width, e.g. c(2,1,1) refers to 2:1:1. Default is NULL for equal column width.}

\item{border_left}{A character vector for left border type. Default is "single". All possible input can be found in \code{r2rtf:::border_type()}.}

\item{border_right}{A character vector for right border type. Default is "single". All possible input can be found in \code{r2rtf:::border_type()}.}

\item{border_top}{A character vector for top border type. Default is NULL. If it is the first row in a table for this page,
the border is set to "double" otherwise the border is set to "single".}

\item{border_bottom}{A character vector for bottom border type. Default is NULL for no border. All possible input can be found in \code{r2rtf:::border_type()}.}

\item{border_color_left}{A character vector for left border color. Default is NULL for black. All possible input can be found in \code{grDevices::colors()}.}

\item{border_color_right}{A character vector for right border color. Default is NULL for black. All possible input can be found in \code{grDevices::colors()}.}

\item{border_color_top}{A character vector for top border color. Default is NULL for black. All possible input can be found in \code{grDevices::colors()}.}

\item{border_color_bottom}{A character vector for bottom border color. Default is NULL for black. All possible input can be found in \code{grDevices::colors()}.}

\item{border_width}{A numeric value for border width in twips. Default is 15.}

\item{cell_height}{A numeric value for height of cell in inch. Default is 0.15 inch.}

\item{cell_justification}{A character vector for justification of cell. Default is "c" for center justification. All possible input can be found in \code{r2rtf:::justification()}.}

\item{text_font}{A numeric value for text font type. Default is 1 for Times New Roman. All possible input can be found in \code{r2rtf:::font_type()}.}

\item{text_format}{A character vector for  text format. Default is NULL for normal format. Multiple format type can be assigned, e.g. "bi" will
give the text both bold and italics format. All possible input can be found in \code{r2rtf:::font_format()}.}

\item{text_font_size}{A numeric value for text font size. Default is 9.}

\item{text_color}{A character vector for text color. Default is NULL for black. All possible input can be found in \code{grDevices::colors()}.}

\item{text_background_color}{A character vector for text background color. Default is NULL for black. All possible input can be found in \code{grDevices::colors()}.}

\item{text_justification}{A character vector for text justification. Default is "c" for center justification. All possible input can be found in \code{r2rtf:::justification()}.}

\item{text_space_before}{A numeric value for line space before text in twips. Default is 15.}

\item{text_space_after}{A numeric value for line space after text in twips. Default is 15.}
}
\value{
The same data frame \code{tbl} with additional attributes for table column header.
}
\description{
Add Column Header Attributes to Table
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Input checks using \code{check_args()}, \code{match_arg()} and \code{stopifnot()}. The required argument is \code{tbl}, i.e. A data frame must define by \code{tbl}.
   \item Set default page attributes and register use_color attribute.
   \item Define column header attributes of \code{tbl} based on the input.
   \item Return \code{tbl}.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
library(dplyr) # required to run examples
data(tbl_1)
tbl_1 \%>\%
  rtf_colheader(colheader = "Treatment | N | Mean (SD) | N | Mean (SD) | N |
                  Mean (SD) | LS Mean (95\% CI)\\\\dagger",
                text_format = c("b", "", "u", "", "u", "","u", "i") ) \%>\%
  attr("rtf_colheader")

}
