% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_convert_format.R
\name{rtf_convert_format}
\alias{rtf_convert_format}
\title{Convert RTF file to Other Format}
\usage{
rtf_convert_format(
  input,
  output_file = NULL,
  output_dir = ".",
  format = "pdf",
  overwrite = FALSE
)
}
\arguments{
\item{input}{A vector of file paths for the input file to be converted.}

\item{output_file}{A vector of filename for the output file.
Default is the same filename for input.}

\item{output_dir}{The output directory for the converted \code{output_dir}.}

\item{format}{Converted file format extension. Currently support "pdf" or "docx"}

\item{overwrite}{logical; should existing destination files be overwritten?}
}
\value{
A vector of file paths for the converted files.
}
\description{
This is an experimental function.
}
\details{
Convert RTF files to PDF or DOCX files. Require \code{libreoffice}.
}
