% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_encode.R
\name{rtf_encode}
\alias{rtf_encode}
\title{Render to RTF Encoding}
\usage{
rtf_encode(
  tbl,
  type = "table",
  page_title = "all",
  page_footnote = "last",
  page_source = "last"
)
}
\arguments{
\item{tbl}{a data frame for table or a list of binary string for figure}

\item{type}{the type of input, default is table.}

\item{page_title}{a character of title displaying location. Default is "all" for all pages.
Possible values are "first", "last" and "all".}

\item{page_footnote}{a character of title displaying location. Default is "last" for all pages.
Possible values are "first", "last" and "all".}

\item{page_source}{a character of title displaying location. Default is "last" for all pages.
Possible values are "first", "last" and "all".}
}
\value{
\preformatted{For \code{rtf_encode}, a vector of RTF code.
For \code{write_rtf}, no return value.
}
}
\description{
This function extracts table/figure attributes and render to RTF encoding that is ready to save
to an RTF file.
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Input check for type ("table" or "figure").
   \item Input check for title, footnote and source position ("all", "first" or "last").
   \item If type is "table" and class is data.frame then run \code{rtf_encode_table(tbl)}.
   \item If type is "table" and class is list then run \code{rtf_encode_list(tbl)}.
   \item If type is "figure" then run \code{rtf_encode_figure(tbl)}.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{

library(dplyr)  # required to run examples

# Example 1
  head(iris) \%>\%
    rtf_body() \%>\%
    rtf_encode() \%>\%
    write_rtf(file = file.path(tempdir(), "table1.rtf"))

# Example 2
\dontrun{
  library(dplyr) # required to run examples
  file <- file.path(tempdir(), "figure1.png")
  png(file)
  plot(1:10)
  dev.off()

  # Read in PNG file in binary format
  rtf_read_png(file) \%>\% rtf_figure() \%>\%
    rtf_encode(type = "figure") \%>\%
    write_rtf(file = file.path(tempdir(), "figure1.rtf"))
}
# Example 3

## convert tbl_1 to the table body. Add title, subtitle, two table
## headers, and footnotes to the table body.
  data(tbl_1)
  data(tbl_2)
  data(tbl_3)
  ## convert tbl_2 to the table body. Add a table column header to table body.
  t2 <- tbl_2 \%>\%
    rtf_colheader(colheader = "Pairwise Comparison |
                   Difference in LS Mean(95\% CI)\\\\dagger | p-Value",
                  text_justification = c("l","c","c")) \%>\%
    rtf_body(col_rel_width = c(8,7,5),
             text_justification = c("l","c","c"),
             last_row = FALSE);

  # concatenate a list of table and save to an RTF file
  t2 \%>\% rtf_encode() \%>\% write_rtf(file.path(tempdir(), "table2.rtf"))

}
