% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_encode_list.R
\name{rtf_encode_list}
\alias{rtf_encode_list}
\title{Render List to RTF Encoding}
\usage{
rtf_encode_list(tbl)
}
\arguments{
\item{tbl}{A data frame for table or a list of binary string for figure.}
}
\description{
Render List to RTF Encoding
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect color attributes from \code{tbl} object.
   \item Initiate RTF using \code{as_rtf_init()}, \code{as_rtf_font()} and color syntax obtained from previous step.
   \item Translate all \code{tbl} attributes into RTF syntax.
   \item Combine all components into a single RTF code string.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

