% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_encode_table.R
\name{rtf_encode_table}
\alias{rtf_encode_table}
\title{Render Table to RTF encoding}
\usage{
rtf_encode_table(
  tbl,
  page_title = "all",
  page_footnote = "last",
  page_source = "last"
)
}
\arguments{
\item{tbl}{a data frame for table or a list of binary string for figure}

\item{page_title}{a character of title displaying location. Default is "all" for all pages.
Possible values are "first", "last" and "all".}

\item{page_footnote}{a character of title displaying location. Default is "last" for all pages.
Possible values are "first", "last" and "all".}

\item{page_source}{a character of title displaying location. Default is "last" for all pages.
Possible values are "first", "last" and "all".}
}
\description{
Render Table to RTF encoding
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Initiate RTF using \code{as_rtf_init()}, \code{as_rtf_font()} and \code{as_rtf_color()}.
   \item Define page, margin, header, footnote, source and new_page in RTF syntax.
   \item Define column header, first border and last border type in RTF syntax.
   \item Check whether footnote and source will be displayed as table if they exist.
   \item Define table content in RTF syntax.
   \item Get page title display location ("all", "first", "last") from arg input and display it in page accordinly.
   \item Get page footnote display location ("all", "first", "last") from arg input and display it in page accordinly.
   \item Get page source display location ("all", "first", "last") from arg input and display it in page accordinly.
   \item Translate all \code{tbl} attributes into RTF syntax.
   \item Combine all components into a single RTF code string.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

