% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_group_by_enhance.R
\name{rtf_group_by_enhance}
\alias{rtf_group_by_enhance}
\title{Remove Duplicate Records}
\usage{
rtf_group_by_enhance(tbl, group_by, page_index)
}
\arguments{
\item{tbl}{A data frame.}

\item{group_by}{A character vector of variable names in \code{tbl}.}

\item{page_index}{A numeric vector of page index.}
}
\value{
Return \code{tbl}.
}
\description{
Remove Duplicate Records
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define \code{id} variable to split data frame.
   \item Remove duplicate records within each splitted data frame.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

