% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_pageby.R
\name{rtf_pageby}
\alias{rtf_pageby}
\title{add table body page by attributes to the table}
\usage{
rtf_pageby(tbl, page_by, new_page, pageby_header)
}
\arguments{
\item{tbl}{A data frame.}

\item{page_by}{Column names in a character vector to group by table in sections.
Default is NULL.}

\item{new_page}{A boolean value to indicate whether to separate grouped table into pages
by sections. Default is FALSE.}

\item{pageby_header}{A boolean value to display pageby header at the beginning of each page.}
}
\description{
add table body page by attributes to the table
}
