% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_figure.R
\name{rtf_read_png}
\alias{rtf_read_png}
\title{Read PNG figures into Binary Files}
\usage{
rtf_read_png(file)
}
\arguments{
\item{file}{a vector of PNG file path}
}
\description{
Read PNG figures into Binary Files
}
\examples{
\dontrun{
  file <- file.path(tempdir(), "figure1.png")
  png(file)
  plot(1:10)
  dev.off()

  # Read in PNG file in binary format
  rtf_read_png(file)
}

}
