% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_span_row.R
\name{rtf_span_row}
\alias{rtf_span_row}
\title{Add Horizontal Span Row Attributes to Table}
\usage{
rtf_span_row(tbl, span_row)
}
\arguments{
\item{tbl}{A data frame.}

\item{span_row}{A logical vector of length \code{nrow(tbl)} indicating
which rows should span all columns, or an integer vector of row indices.}
}
\value{
the same data frame \code{tbl} with additional attributes for horizontal span rows
}
\description{
Add Horizontal Span Row Attributes to Table
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Validate that \code{tbl} has body attributes from \code{rtf_body()}.
   \item Normalize \code{span_row} to a logical vector of length \code{nrow(tbl)}.
   \item Set the \code{"rtf_span_row"} attribute on \code{tbl}.
   \item Return \code{tbl}.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
library(dplyr) # required to run examples
data(r2rtf_tbl1)
r2rtf_tbl1 \%>\%
  rtf_body() \%>\%
  rtf_span_row(span_row = c(rep(TRUE, 2), rep(FALSE, nrow(r2rtf_tbl1) - 2))) \%>\%
  attr("rtf_span_row")

}
