% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_margin.R
\name{set_margin}
\alias{set_margin}
\title{Define Margin Type}
\usage{
set_margin(doctype, orientation)
}
\arguments{
\item{doctype}{doctype in 'csr', 'wma', 'wmm' or 'narrow'}

\item{orientation}{Orientation in 'portrait' or 'landscape'.}
}
\description{
Define Margin Type
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define document margin by assigning margin values for left, right, top, bottom, header and footer.
   \item Define document orientation.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

