% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{spacing}
\alias{spacing}
\title{RTF Paragraph Spacing Dictionary}
\usage{
spacing()
}
\description{
RTF Paragraph Spacing Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect most commonly used paragraph spacing (single-space, double-space, and 1.5-space).
   \item Define the paragraph spacing type in 1, 2, 1.5.
   \item Create a mapping between paragraph spacing and their RTF code.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

