% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_2}
\alias{tbl_2}
\title{Between Group Results from an ANCOVA Model}
\format{A data frame with 1 row and 3 variables.}
\usage{
tbl_2
}
\description{
A dataset containing between group results from an ANCOVA model.
}
\keyword{datasets}
