% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unicode_latex}
\alias{unicode_latex}
\title{Dictionary of Unicode and Latex Code}
\format{
A data frame with 681 rows and 3 variables.
\describe{
\item{unicode}{unicode, UTF-8 code}
\item{latex}{latex, latex code}
\item{int}{int, Converted integer of the UTF-8 code}
}
}
\source{
\url{http://milde.users.sourceforge.net/LUCR/Math/data/unimathsymbols.txt}
}
\usage{
unicode_latex
}
\description{
A dataset containing the mapping between unicode and latex code.
}
\keyword{datasets}
