% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_convert_format.R
\name{update_cellx}
\alias{update_cellx}
\title{Update RTF border twips}
\usage{
update_cellx(x, tolerance = 5)
}
\arguments{
\item{x}{a character vector of RTF encoding.}

\item{tolerance}{tolerance of the difference for \verb{\\\\cellx}.}
}
\value{
a character vector of RTF encoding.
}
\description{
Update RTF border twips
}
\keyword{internal}
