% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{utf8Tortf}
\alias{utf8Tortf}
\title{Convert a UTF-8 Encoded Character String to a RTF Encoded String}
\usage{
utf8Tortf(text)
}
\arguments{
\item{text}{A string to be converted.

If the unicode of a character is 255 or under (including all character on a keyboard), the character is as is.
If the unicode of a character is larger than 255, the character will be encoded.}
}
\description{
Convert a UTF-8 Encoded Character String to a RTF Encoded String
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define rules for character by setting 255 as cutoff.
   \item If the unicode of a character is 255 or under (including all character on a keyboard), the character is as is.
   \item If the unicode of a character is larger than 255, the character will be encoded.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\references{
Burke, S. M. (2003). RTF Pocket Guide. " O'Reilly Media, Inc.".
}
