% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vertical_justification}
\alias{vertical_justification}
\title{RTF Text Vertical Justification Dictionary}
\usage{
vertical_justification()
}
\description{
RTF Text Vertical Justification Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect most commonly used vertical alignments for texts or rows (top and bottom).
   \item Create a mapping between justifications and their RTF code.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

