% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rtf.R
\name{write_rtf}
\alias{write_rtf}
\title{Write an RTF Table or Figure to an RTF File}
\usage{
write_rtf(rtf, file)
}
\arguments{
\item{rtf}{a character rtf encoding string rendered by \code{rtf_encode()}}

\item{file}{a character string naming a file to save rtf file.}
}
\description{
The write_rtf function writes rtf encoding string to an .rtf file
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Export a single RTF string into an file using \code{write} function.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

