% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rtf.R
\name{write_rtf_para}
\alias{write_rtf_para}
\title{Write a Paragraph to an RTF File}
\usage{
write_rtf_para(rtf_body, file)
}
\arguments{
\item{rtf_body}{rtf code for text paragraph, obtained using \code{rtf_paragraph(text,...)} function}

\item{file}{file name to save rtf text paragraph, eg. filename.rtf}
}
\description{
Write a Paragraph to an RTF File
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define table color using \code{color_table()} and translate in RTF syntax.
   \item Initiate rtf using \code{as_rtf_init()} and \code{as_rtf_font()}.
   \item Combine the text with other components into a single RTF code string.
   \item Output the paragraph into a file.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

