% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{font_type}
\alias{font_type}
\title{RTF Text Font Dictionary}
\usage{
font_type()
}
\description{
RTF Text Font Dictionary
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Collect most commonly used fonts (Times New Roman, Times New Roman Greek, and Arial Greek, etc.).
   \item Define font types from 1 to 10.
   \item Define font styles.
   \item Create a mapping between font types and their RTF code.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

