% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HAMD17}
\alias{HAMD17}
\title{HAMD17}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 831 rows and 6 columns.}
\usage{
HAMD17
}
\description{
HAMD17
}
\keyword{datasets}
