% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2rtf_HAMD17}
\alias{r2rtf_HAMD17}
\title{An Efficacy Clinical Trial Data to Evaluate a Drug to Reduce Lower Back Pain}
\format{
A data frame with 831 rows and 6 variables.
}
\source{
Drug Information Association (DIA) scientific working group on missing data
}
\usage{
r2rtf_HAMD17
}
\description{
A dataset prepared by the Drug Information Association scientific working group
to investigate a drug to reduce lower back pain.
}
\details{
This dataset was prepared by the Drug Information Association (DIA) scientific working group
to investigate methods for handling missing data. It contains data from an efficacy clinical trial
evaluating a drug to reduce lower back pain.
}
\keyword{datasets}
