% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_assemble.R
\name{rtf_assemble}
\alias{rtf_assemble}
\title{Assemble Multiple RTF Table Listing and Figure Into One Document}
\usage{
rtf_assemble(input, output, landscape = FALSE, use_officer = TRUE)
}
\arguments{
\item{input}{Character vector of file path.}

\item{output}{Character string to the output file path.}

\item{landscape}{Logical vector to determine whether to
display files as portrait or landscape. If \code{use_officer = FALSE}, only singular value is allowed.}

\item{use_officer}{Logical value to determine whether the package \code{officer}
should be used to assemble RTF table, listing and figure.}
}
\description{
The function asseble multiple RTF table, listing, and figures into one document
as RTF file or Microsft Word in \code{docx} format.
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Transfer files to toggle fields format in Word
    \item Insert into Word file using officer
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
input <- list.files("vignettes/rtf", pattern = "*.rtf", full.names = TRUE)
output <- tempfile(fileext = ".rtf")

rtf_assemble(
  input = input,
  output = output,
  use_officer = FALSE
)

}
