% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rtf.R
\name{rtf_encode}
\alias{rtf_encode}
\alias{as_rtf}
\title{Render to RTF Encoding}
\usage{
rtf_encode(tbl, type = "table")

as_rtf(tbl, type = "table")
}
\arguments{
\item{tbl}{a data frame for table or a list of binary string for figure}

\item{type}{the type of input, default is table.}
}
\description{
This function extracts table/figure attributes and render to RTF encoding that is ready to save
to an RTF file.
}
