% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_figure.R
\name{rtf_figure}
\alias{rtf_figure}
\title{Add Figure Attrbuties}
\usage{
rtf_figure(gt_tbl, page_width = 8.5, page_height = 11,
  orientation = "portrait", doctype = "wma", fig_width = 5,
  fig_height = 5)
}
\arguments{
\item{gt_tbl}{A data frame}

\item{page_width}{page width in inches}

\item{page_height}{page height in inches}

\item{orientation}{Orientation in 'portrait' or 'landscape'}

\item{doctype}{doctype in 'csr', 'wma', or 'wmm'}

\item{fig_width}{the width of figures in inch}

\item{fig_height}{the height of figures in inch}
}
\description{
Add Figure Attrbuties
}
