% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_pageby.R
\name{rtf_pageby}
\alias{rtf_pageby}
\title{Add Table Body Pageby Attributes to Table}
\usage{
rtf_pageby(
  tbl,
  page_by = NULL,
  new_page = FALSE,
  pageby_header = TRUE,
  pageby_row = "column"
)
}
\arguments{
\item{tbl}{A data frame.}

\item{page_by}{Column names in a character vector to group by table in sections.
Default is NULL.}

\item{new_page}{A boolean value to indicate whether to separate grouped table into pages
by sections. Default is FALSE.}

\item{pageby_header}{A boolean value to display pageby header at the beginning of each page.}

\item{pageby_row}{A character vector of location of page_by variable. Possible input are 'column'
or 'first_row'. Default is 'column'.}
}
\description{
Add Table Body Pageby Attributes to Table
}
