% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_read.R
\name{rtf_read_png}
\alias{rtf_read_png}
\title{Read PNG Figures into Binary Files}
\usage{
rtf_read_png(file)
}
\arguments{
\item{file}{A character vector of PNG file paths.}
}
\value{
a list of binary data vector returned by \code{readBin}
}
\description{
Read PNG Figures into Binary Files
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Deprecated: rtf_read_png. Use rtf_read_figure instead
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

