% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_table_content.R
\name{rtf_table_content}
\alias{rtf_table_content}
\title{Create RTF Table Body Encode}
\usage{
rtf_table_content(
  db,
  col_total_width = attr(db, "page")$col_width,
  use_border_bottom = FALSE,
  border_left = attr(db, "border_left"),
  border_right = attr(db, "border_right"),
  border_top = attr(db, "border_top"),
  border_bottom = attr(db, "border_bottom"),
  border_color_left = attr(db, "border_color_left"),
  border_color_right = attr(db, "border_color_right"),
  border_color_top = attr(db, "border_color_top"),
  border_color_bottom = attr(db, "border_color_bottom"),
  border_width = attr(db, "border_width"),
  col_rel_width = attr(db, "col_rel_width"),
  cell_height = attr(db, "cell_height"),
  cell_justification = attr(db, "cell_justification"),
  text_font = attr(db, "text_font"),
  text_format = attr(db, "text_format"),
  text_color = attr(db, "text_color"),
  text_background_color = attr(db, "text_background_color"),
  text_justification = attr(db, "text_justification"),
  text_font_size = attr(db, "text_font_size"),
  text_space_before = attr(db, "text_space_before"),
  text_space_after = attr(db, "text_space_after"),
  text_convert = attr(db, "text_convert")
)
}
\arguments{
\item{db}{a data frame}

\item{col_total_width}{column total width for the table}

\item{use_border_bottom}{logical value of using the bottom border}

\item{border_left}{left border type}

\item{border_right}{right border type}

\item{border_top}{top border type}

\item{border_bottom}{bottom border type}

\item{border_color_left}{left border color}

\item{border_color_right}{right border color}

\item{border_color_top}{top border color}

\item{border_color_bottom}{bottom border color}

\item{border_width}{border width in twips}

\item{col_rel_width}{column relative width in a vector eg. c(2,1,1) refers to 2:1;1}

\item{cell_height}{height for cell in twips}

\item{cell_justification}{justification for cell}

\item{text_font}{text font type}

\item{text_format}{text format}

\item{text_color}{text color}

\item{text_background_color}{text background color}

\item{text_justification}{justification for text}

\item{text_font_size}{text font size}

\item{text_space_before}{line space before text}

\item{text_space_after}{line space after text}

\item{text_convert}{a logical value to convert special characters. Default is TRUE.}
}
\description{
Create RTF Table Body Encode
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Define table begin and end in RTF syntax.
   \item Define cell justification using `justification()`, then covert the cell from inch to twip using `inch_to_twip()` in RTF syntax.
   \item Define cell border type using `border_type()` and cell border width in RTF syntax.
   \item Define cell border color using `color_table()` in RTF syntax.
   \item Define cell background color using input variable text_background_color in RTF syntax.
   \item Define cell size using `cell_size()` in RTF syntax.
   \item Combine cell component attributes into a single code string.
   \item Define cell content in encoded RTF syntax.
   \item Check if cell content format is a valid value.
   \item Combine cell content and content component attributes into a single code string.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

