% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/internal_datasets.R
\docType{data}
\name{unicode_latex}
\alias{unicode_latex}
\title{unicode_latex}
\format{An object of class \code{data.frame} with 681 rows and 3 columns.}
\source{
\url{http://milde.users.sourceforge.net/LUCR/Math/data/unimathsymbols.txt}
}
\usage{
unicode_latex

unicode_latex
}
\description{
A dataset containing the unicode, character and latex command.
}
\keyword{datasets}
