% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rtf.R
\name{write_rtf}
\alias{write_rtf}
\title{Write an RTF table to .rtf file}
\usage{
write_rtf(rtf, file)
}
\arguments{
\item{rtf}{rtf encoing string rendered by \code{as_rtf()}}

\item{file}{File name to write the output RTF table.}
}
\description{
The write_rtf function writes rtf encoding string to an .rtf file
}
